feat: run commands in cli
