+++ src/app.js
@@ -1,6 +1,6 @@
 import './filesystem.js';
 
-import { COMMANDS, COMMANDS_RUN } from './common/command.js';
+import { COMMANDS, COMMANDS_RUN } from './common/commands.js';
 import readlinePromises from 'node:readline/promises';
 
 
@@ -11,17 +11,27 @@ const rl = readlinePromises.createInterface({
   output: stdout,
 })
 
-const runApp = () => {
+const runApp = async () => {
     rl.on('line', (line) => {
       const query = line.toString().trim();
-      if (query === COMMANDS.EXIT) COMMANDS_RUN.EXIT();
-      console.log(`> ${query}`)
+      const commandKeys = Object.keys(COMMANDS)
+      let commandMatched = false;
+
+      commandKeys.find(key => {
+          if (query === COMMANDS[key]) {
+          stdout.write(`> ${COMMANDS_RUN[key]()}\n`)
+          commandMatched = true;
+        }
+      })
+
+      if(!commandMatched) {
+        stdout.write("> Please enter a valid command\n");
+      }
     })
-    
+
     rl.on('SIGINT', () => {
       COMMANDS_RUN.EXIT();
     })
-
 }
 
 runApp();
