--- src/common/command.js
@@ -1,13 +1,0 @@
-import { exit } from '../util/exit.js';
-import { userName } from '../util/user-name.js';
-
-const name = userName();
-
-export const COMMANDS = {
-  EXIT: '.exit',
-}
-
-export const COMMANDS_RUN = {
-  EXIT: () => exit(name),
-}
-
